fix：修复权限问题

+++ routes/admin/user_admin.py
@@ -52,24 +52,27 @@ async def forbid_user(
 ):
     if (
         not access_token
-        and not username
-        and not isinstance(minutes, int)
-        and minutes < 0
+        or not username
+        or not isinstance(minutes, int)
+        or not minutes < 0
     ):
-        username_use = await login_util.get_user_from_jwt(access_token)
-        if not (
-            username_use
-            and await login_util.is_login(access_token, username_use)
-            and await admin_util.is_admin(username_use)
-        ):
-            return JSONResponse(content={"error": "参数错误"}, status_code=400)
+        return JSONResponse(content={"error": "参数错误"}, status_code=400)
 
-    result = await admin_util.forbid_user(username, minutes)
-    return (
-        JSONResponse(content={"status": "success"}, status_code=200)
-        if result
-        else JSONResponse(content={"status": "fail"}, status_code=400)
-    )
+    username_use = await login_util.get_user_from_jwt(access_token)
+    if (
+        username_use
+        and await login_util.is_login(access_token, username_use)
+        and await admin_util.is_admin(username_use)
+    ):
+        # 如果用户不是超级管理员，不能禁言其他管理员
+        if await admin_util.is_admin(username) and username_use != "admin":
+            return JSONResponse(content={"status": "fail"}, status_code=400)
+
+        result = await admin_util.forbid_user(username, minutes)
+        if result:
+            return JSONResponse(content={"status": "success"}, status_code=200)
+
+    return JSONResponse(content={"status": "fail"}, status_code=400)
 
 
 # 解除禁言
@@ -78,21 +81,24 @@ async def unforbid_user(
     access_token: Optional[str] = Cookie(None),
     username: str = Query(None),
 ):
-    if not access_token and not username:
-        username_use = await login_util.get_user_from_jwt(access_token)
-        if not (
-            username_use
-            and await login_util.is_login(access_token, username_use)
-            and await admin_util.is_admin(username_use)
-        ):
-            return JSONResponse(content={"error": "参数错误"}, status_code=400)
+    if not access_token or not username:
+        return JSONResponse(content={"error": "参数错误"}, status_code=400)
 
-    result = await admin_util.unforbid_user(username)
-    return (
-        JSONResponse(content={"status": "success"}, status_code=200)
-        if result
-        else JSONResponse(content={"status": "fail"}, status_code=400)
-    )
+    username_use = await login_util.get_user_from_jwt(access_token)
+    if (
+        username_use
+        and await login_util.is_login(access_token, username_use)
+        and await admin_util.is_admin(username_use)
+    ):
+        # 如果用户不是超级管理员，不能解除其他管理员的禁言
+        if await admin_util.is_admin(username) and username_use != "admin":
+            return JSONResponse(content={"status": "fail"}, status_code=400)
+
+        result = await admin_util.unforbid_user(username)
+        if result:
+            return JSONResponse(content={"status": "success"}, status_code=200)
+
+    return JSONResponse(content={"status": "fail"}, status_code=400)
 
 
 # 为用户添加管路权限
@@ -102,21 +108,16 @@ async def add_admin(
     username: str = Query(None),  # 用户名
     days: int = Query(None),  # 管理员权限天数
 ):
-    if not access_token and not username and not days and not isinstance(days, int):
-        username_use = await login_util.get_user_from_jwt(access_token)
-        if not (
-            username_use
-            and await login_util.is_login(access_token, username_use)
-            and await admin_util.is_admin(username_use)
-        ):
-            return JSONResponse(content={"error": "参数错误"}, status_code=400)
+    if not access_token or not username or not days or not isinstance(days, int):
+        return JSONResponse(content={"error": "参数错误"}, status_code=400)
 
-    result = await admin_util.add_admin(username, days)
-    return (
-        JSONResponse(content={"status": "success"}, status_code=200)
-        if result
-        else JSONResponse(content={"status": "fail"}, status_code=400)
-    )
+    username_use = await login_util.get_user_from_jwt(access_token)
+    if username_use == "admin":
+        result = await admin_util.add_admin(username, days)
+        if result:
+            return JSONResponse(content={"status": "success"}, status_code=200)
+
+    return JSONResponse(content={"status": "fail"}, status_code=400)
 
 
 # 删除用户的管理员权限
@@ -125,18 +126,13 @@ async def delete_admin(
     access_token: Optional[str] = Cookie(None),
     username: str = Query(None),  # 用户名
 ):
-    if not access_token and not username:
-        username_use = await login_util.get_user_from_jwt(access_token)
-        if not (
-            username_use
-            and await login_util.is_login(access_token, username_use)
-            and await admin_util.is_admin(username_use)
-        ):
-            return JSONResponse(content={"error": "参数错误"}, status_code=400)
+    if access_token and username:
+        return JSONResponse(content={"error": "参数错误"}, status_code=400)
 
-    result = await admin_util.delete_admin(username)
-    return (
-        JSONResponse(content={"status": "success"}, status_code=200)
-        if result
-        else JSONResponse(content={"status": "fail"}, status_code=400)
-    )
+    username_use = await login_util.get_user_from_jwt(access_token)
+    if username_use == "admin":
+        result = await admin_util.delete_admin(username)
+        if result:
+            return JSONResponse(content={"status": "success"}, status_code=200)
+
+    return JSONResponse(content={"status": "fail"}, status_code=400)
